Correct logic so that enpassantflag is set to false everytime a piece moves.
